Fixes for typos in code and docs. And a new package for api files
